Split Augury into clear major subsystems and other misc refactorings (#40)
We now have 3 major pieces that are very clear and obvious due to these changes: An EventDispatcher that dispatches probes events, a History Manager and a Plugin Manager. Basically works like this:

Plugins ---> Augury/History <---- Event Dispatcher (Probes)

More refactorings are necessary to keep cleaning things up and adding more types, but this is a good step in the right direction I think.

Removed Commands - they are not needed and oddly wired into the event dispatcher just to take part in reactions, but the logic only really depended on the history and nothing else. Removing them eliminated all the coupling with the other major managers/services of augury.

Removed old event dispatcher and renamed ProbeManager to EventDispatcher. These two classes should be combined and were honestly pretty redundant

Removed enhancers. They only applied to 1 event from 1 probe each, and there was no reason the probe itself couldn't just handle it. If we need a way to apply transformations to multiple events coming from different types of probes, we can add it later when we actually have a requirement to do so. I felt like the current design was over-engineered a bit for what we really needed.

Removed reactions. They are no longer needed once commands were removed. Non-command events didn't seem use reactions at all.

Renamed Augury Event types for added consistency. It's actually quite possible that AuguryEventSource isn't even needed anymore.

Drastically simplified references within AuguryCore

Plugins now have a reference to Augury itself, which enables them to interact with the history or create live channels. Now AuguryCore is acting as a Facade to other applications rather than just storing references of other dependencies

Simplified Channels a bit - removed delegate as it wasn't needed.

Renamed all the "EID" fields and made sure all of the plugin/execution timeline interfaces had the same event id and timestamp names to avoid the immense amount of confusion. We really do need add types to all of these Reducers and different types for the AuguryBridgeMessages - it's not very good yet but we can do that in another PR.

Removed NgDebugProbe - it wasn't even doing what the other probes did... it was more like a class filled with helper functions that Commands/Reactions required... and this was not the right place for it. In fact, Enhancers had a reference to the probe manager as a result of this coupling, and it's really good that we removed all of the tangling now.

Changed AuguryEvent interface to a class. We can do more refactoring here over time, but it's much simpler to grok now.

Decoupled bridge reference from the plugin controller/window. Made the bridge into a singleton in case there are multiple plugins active at the same time. It's maybe a good idea to move the bridge into AuguryCore probably.